+++ integer/src/bits.rs
@@ -9,6 +9,8 @@ use core::{
     ops::{BitAnd, BitAndAssign, BitOr, BitOrAssign, BitXor, BitXorAssign, Not},
 };
 
+// Ops for UBig
+
 impl UBig {
     /// Set the `n`-th bit, n starts from 0.
     ///
@@ -206,6 +208,8 @@ impl PowerOfTwo for UBig {
         UBig(self.into_repr().next_power_of_two())
     }
 }
+
+// Ops for IBig
 
 impl IBig {
     /// Returns the number of trailing zeros in the two's complement binary representation.
@@ -1052,6 +1056,59 @@ helper_macros::impl_binop_assign_by_taking!(impl BitAndAssign<IBig> for IBig, bi
 helper_macros::impl_binop_assign_by_taking!(impl BitOrAssign<IBig> for IBig, bitor_assign, bitor);
 helper_macros::impl_binop_assign_by_taking!(impl BitXorAssign<IBig> for IBig, bitxor_assign, bitxor);
 
+// Ops between UBig & IBig
+
+macro_rules! impl_ubig_ibig_bitand {
+    ($sign0:ident, $mag0:ident, $sign1:ident, $mag1:ident) => {{
+        debug_assert_eq!($sign0, Positive);
+        match $sign1 {
+            Positive => UBig($mag0.bitand($mag1)),
+            Negative => UBig($mag0.and_not($mag1.sub_one().into_typed())),
+        }
+    }};
+}
+macro_rules! impl_ibig_ubig_bitand {
+    ($sign0:ident, $mag0:ident, $sign1:ident, $mag1:ident) => {{
+        debug_assert_eq!($sign1, Positive);
+        match $sign0 {
+            Positive => UBig($mag1.bitand($mag0)),
+            Negative => UBig($mag1.and_not($mag0.sub_one().into_typed())),
+        }
+    }};
+}
+helper_macros::forward_ubig_ibig_binop_to_repr!(
+    impl BitAnd,
+    bitand,
+    Output = UBig,
+    impl_ubig_ibig_bitand
+);
+helper_macros::forward_ubig_ibig_binop_to_repr!(impl BitOr, bitor, Output = IBig, impl_ibig_bitor);
+helper_macros::forward_ubig_ibig_binop_to_repr!(
+    impl BitXor,
+    bitxor,
+    Output = IBig,
+    impl_ibig_bitxor
+);
+helper_macros::impl_binop_assign_by_taking!(impl BitAndAssign<IBig> for UBig, bitand_assign, bitand);
+helper_macros::forward_ibig_ubig_binop_to_repr!(
+    impl BitAnd,
+    bitand,
+    Output = UBig,
+    impl_ibig_ubig_bitand
+);
+helper_macros::forward_ibig_ubig_binop_to_repr!(impl BitOr, bitor, Output = IBig, impl_ibig_bitor);
+helper_macros::forward_ibig_ubig_binop_to_repr!(
+    impl BitXor,
+    bitxor,
+    Output = IBig,
+    impl_ibig_bitxor
+);
+helper_macros::impl_binop_assign_by_taking!(impl BitAndAssign<UBig> for IBig, bitand_assign, bitand);
+helper_macros::impl_binop_assign_by_taking!(impl BitOrAssign<UBig> for IBig, bitor_assign, bitor);
+helper_macros::impl_binop_assign_by_taking!(impl BitXorAssign<UBig> for IBig, bitxor_assign, bitxor);
+
+// Ops with primitives
+
 macro_rules! impl_bit_ops_primitive_with_ubig {
     ($($t:ty)*) => {$(
         helper_macros::impl_commutative_binop_with_primitive!(impl BitAnd<$t> for UBig, bitand -> $t);
@@ -1064,7 +1121,19 @@ macro_rules! impl_bit_ops_primitive_with_ubig {
 }
 impl_bit_ops_primitive_with_ubig!(u8 u16 u32 u64 u128 usize);
 
-macro_rules! impl_bit_ops_primitive_with_ibig {
+macro_rules! impl_bit_ops_unsigned_with_ibig {
+    ($($t:ty)*) => {$(
+        helper_macros::impl_commutative_binop_with_primitive!(impl BitAnd<$t> for IBig, bitand -> $t);
+        helper_macros::impl_commutative_binop_with_primitive!(impl BitOr<$t> for IBig, bitor);
+        helper_macros::impl_commutative_binop_with_primitive!(impl BitXor<$t> for IBig, bitxor);
+        helper_macros::impl_binop_assign_with_primitive!(impl BitAndAssign<$t> for IBig, bitand_assign);
+        helper_macros::impl_binop_assign_with_primitive!(impl BitOrAssign<$t> for IBig, bitor_assign);
+        helper_macros::impl_binop_assign_with_primitive!(impl BitXorAssign<$t> for IBig, bitxor_assign);
+    )*};
+}
+impl_bit_ops_unsigned_with_ibig!(u8 u16 u32 u64 u128 usize);
+
+macro_rules! impl_bit_ops_signed_with_ibig {
     ($($t:ty)*) => {$(
         helper_macros::impl_commutative_binop_with_primitive!(impl BitAnd<$t> for IBig, bitand);
         helper_macros::impl_commutative_binop_with_primitive!(impl BitOr<$t> for IBig, bitor);
@@ -1074,7 +1143,7 @@ macro_rules! impl_bit_ops_primitive_with_ibig {
         helper_macros::impl_binop_assign_with_primitive!(impl BitXorAssign<$t> for IBig, bitxor_assign);
     )*};
 }
-impl_bit_ops_primitive_with_ibig!(i8 i16 i32 i64 i128 isize);
+impl_bit_ops_signed_with_ibig!(i8 i16 i32 i64 i128 isize);
 
 #[cfg(test)]
 mod tests {