+++ integer/src/mul_ops.rs
@@ -14,19 +14,8 @@ macro_rules! impl_ibig_mul {
 helper_macros::forward_ibig_binop_to_repr!(impl Mul, mul, Output = IBig, impl_ibig_mul);
 helper_macros::impl_binop_assign_by_taking!(impl MulAssign<IBig> for IBig, mul_assign, mul);
 
-macro_rules! impl_ubig_ibig_mul {
-    ($mag0:ident, $sign1:ident, $mag1:ident) => {
-        IBig($mag0.mul($mag1).with_sign($sign1))
-    };
-}
-helper_macros::forward_ubig_ibig_binop_to_repr!(impl Mul, mul, Output = IBig, impl_ubig_ibig_mul);
-
-macro_rules! impl_ibig_ubig_mul {
-    ($sign0:ident, $mag0:ident, $mag1:ident) => {
-        IBig($mag0.mul($mag1).with_sign($sign0))
-    };
-}
-helper_macros::forward_ibig_ubig_binop_to_repr!(impl Mul, mul, Output = IBig, impl_ibig_ubig_mul);
+helper_macros::forward_ubig_ibig_binop_to_repr!(impl Mul, mul, Output = IBig, impl_ibig_mul);
+helper_macros::forward_ibig_ubig_binop_to_repr!(impl Mul, mul, Output = IBig, impl_ibig_mul);
 helper_macros::impl_binop_assign_by_taking!(impl MulAssign<UBig> for IBig, mul_assign, mul);
 
 // Ops with primitives