+++ integer/tests/div.rs
@@ -297,7 +297,7 @@ fn test_div_rem_ubig_unsigned() {
     let mut x = ubig!(23);
     x /= &10u8;
     assert_eq!(x, ubig!(2));
-    
+
     assert_eq!(23u8 / ubig!(10), 2);
     assert_eq!(23u8 / &ubig!(10), 2);
     assert_eq!(&23u8 / ubig!(10), 2);